Refactor to reduce file size
