+++ 2024-frontend/src/utils.js
@@ -2,3 +2,8 @@
 export const formatError = (data) => {
   return data.$errors.map((e) => e.$message).join(" ")
 }
+
+export const formatNoValue = (value) => {
+  if (value || value === 0) return value
+  return "—"
+}
