+++ 2024-frontend/src/views/WasteMeasurements/index.vue
@@ -1,6 +1,8 @@
 <script setup>
 import { computed, onMounted, ref } from "vue"
+import { formatNoValue } from "@/utils"
 // import BreadcrumbsNav from "@/components/BreadcrumbsNav.vue"
+import MeasurementDetail from "./MeasurementDetail.vue"
 
 const props = defineProps(["canteenUrlComponent"])
 const canteenId = computed(() => props.canteenUrlComponent.split("--")[0])
@@ -53,72 +55,6 @@ const measurementPercentageValues = computed(() => {
 })
 
 const detailsExpandedId = ref("")
-const detailedFields = [
-  {
-    key: "totalMass",
-    label: "Masse totale de gaspillage relevée sur la période de mesure",
-  },
-  // TODO: calculated value - from backend?
-  // {
-  //   label: "Période de mesure de mon gaspillage alimentaire",
-  //   unit: "jours",
-  // },
-  {
-    label: "Nombre de couverts sur la période",
-    unit: "couverts",
-    key: "mealCount",
-  },
-  // TODO: calculated value with tooltip - from backend?
-  // {
-  //   label: "Masse totale de gaspillage sur l'année",
-  //   unit: "kg",
-  // },
-  {
-    heading: "Mesure du gaspillage au niveau des excédents de préparation",
-  },
-  {
-    label: "Masse de gaspillage d'excédents de préparation",
-    key: "preparationTotalMass",
-  },
-  {
-    label: "Masse de gaspillage de denrées comestibles parmi les excédents de préparation",
-    key: "preparationEdibleMass",
-  },
-  {
-    label: "Masse de gaspillage de denrées non comestibles parmi les excédents de préparation",
-    key: "preparationInedibleMass",
-  },
-  {
-    heading: "Mesure du gaspillage au niveau des denrées présentées aux convives mais non servies",
-  },
-  {
-    label: "Masse de gaspillage de denrées présentées aux convives mais non servies",
-    key: "unservedTotalMass",
-  },
-  {
-    label: "Masse de gaspillage de denrées comestibles parmi les denrées présentées aux convives mais non servies",
-    key: "unservedEdibleMass",
-  },
-  {
-    label: "Masse de gaspillage de denrées non comestibles parmi les denrées présentées aux convives mais non servies",
-    key: "unservedInedibleMass",
-  },
-  {
-    heading: "Mesure du gaspillage au niveau du reste assiette",
-  },
-  {
-    label: "Masse de gaspillage pour le reste assiette",
-    key: "leftoversTotalMass",
-  },
-  {
-    label: "Masse de gaspillage de denrées comestibles parmi le reste assiette",
-    key: "leftoversEdibleMass",
-  },
-  {
-    label: "Masse de gaspillage de denrées non comestibles parmi le reste assiette",
-    key: "leftoversInedibleMass",
-  },
-]
 
 onMounted(() => {
   fetch(`/api/v1/canteens/${canteenId.value}/wasteMeasurements`)
@@ -127,11 +63,6 @@ onMounted(() => {
       measurements.value = response
     })
 })
-
-const formatNoValue = (value) => {
-  if (value || value === 0) return value
-  return "—"
-}
 </script>
 
 <template>
@@ -193,16 +124,7 @@ const formatNoValue = (value) => {
             </router-link>
           </div>
         </div>
-        <div v-for="field in detailedFields" :key="field.key" class="detail">
-          <p v-if="field.label" class="fr-mb-1w">{{ field.label }}</p>
-          <h4 v-else-if="field.heading" class="fr-h6">{{ field.heading }}</h4>
-          <p v-if="field.key">
-            <b>
-              {{ formatNoValue(measurement[field.key]) }}
-              {{ field.unit || "kg" }}
-            </b>
-          </p>
-        </div>
+        <MeasurementDetail :measurement="measurement" />
         <router-link class="fr-btn fr-btn--tertiary fr-btn--sm" :to="measurementTunnel">
           <span class="fr-icon-pencil-line fr-icon--sm fr-mr-1w"></span>
           Modifier les données
@@ -254,9 +176,5 @@ p.highlight.brown {
 }
 p.highlight.brown > span {
   background: var(--orange-terre-battue-975-75);
-}
-
-div.detail {
-  color: var(--text-mention-grey);
 }
 </style>
